Task7

+++ Tyuiu.AntonovNS.Sprint2.Task7.V2/Program.cs
@@ -24,9 +24,9 @@ namespace Tyuiu.AntonovNS.Sprint2.Task6.V2
             Console.WriteLine("***************************************************************************");
             double x,y;
             Console.WriteLine("Введите значние X: ");
-            x = Convert.ToInt32(Console.ReadLine());
+            x = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Введите значние Y: ");
-            y = Convert.ToInt32(Console.ReadLine());
+            y = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("***************************************************************************");
             Console.WriteLine("* РЕЗУЛЬТАТ:                                                              *");
             Console.WriteLine("***************************************************************************");
